Fix set_the_flash_matcher to work with Rails 4
The set_the_flash matcher will dup the flash hash for the controller on which it
operates. As #dup does a shallow copy we need to make sure to also copy the
instance variables within the flash hash when we do this. The Rails 4 version
of FlashHash has an extra @discard variable which we have neglected to copy in
this section of code, so this commit makes sure that we do that.